requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
gTTS==2.3.2
idna==3.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
